app/requirements.txt
fastapi
uvicorn[standard]
pillow
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
psycopg[binary]
python-dotenv
pgvector
requests
# Needed for FastAPI file uploads
python-multipart
# OAuth and session management
authlib
python-jose[cryptography]
# For CPU-only PyTorch wheels:
torch
torchvision
